frontend_app/requirements.txt
streamlit
requests
google-cloud-storage
google-auth